Initial Working Version: a very first draft
A LOT of work still needs to be done.

Regarding the .gitignore file - it's just ignoring the hard-coded
test-files and other random files I have lying in my working dir.

Known Issue:
The subsequent executions after the first one on the same .gnucash
file throws the following error a number of times - the repetition
seems to be correlated with the number of changes made in the pre-
vious execution(s?) (No Linebreaks):
```
* 11:49:00  WARN <gnc.backend.xml> Invalid timestamp in data file.
Look for a 'entry:entered' entry with a date of 1969-12-31 or 1970-01-01.
```