Such Modularization. Much Amaze.
Modularized code, a lot. Defined a class to aid further modulariz-
ation. Removed hard-coded values. Turned the script itself into a
module that can be imported, and it's functions used.

Hopefully, this script will soon turn out to be super flexible and
useful to others as well.

Also, added *.pyc to the .gitignore file (now that the script can
be imported as a module).